Consolidate threshold elision logic in nonnegativity analysis
